Split library up into different files
