+++ R.swift Library/Rswift/UIKit/UIViewController+StoryboardSegueIdentifierProtocol.swift
@@ -1,0 +1,16 @@
+//
+//  UIViewController+StoryboardSegueIdentifierProtocol.swift
+//  R.swift Library
+//
+//  Created by Mathijs Kadijk on 06-12-15.
+//  Copyright © 2015 Mathijs Kadijk. All rights reserved.
+//
+
+import Foundation
+import UIKit
+
+public extension UIViewController {
+  public func performSegueWithIdentifier<Identifier: StoryboardSegueIdentifierProtocol>(identifier: Identifier, sender: AnyObject?) {
+    performSegueWithIdentifier(identifier.identifier, sender: sender)
+  }
+}
